Declare Cython-dependency as setup req.
Strictly speacking, this is a lie, as we do not need `Cython` for
setup. But it is a requirement of `kivy`, which is not declared by
`kivy`. Therefore we have to use other tricks to make `pip` and
`setuptools` have Cython installed before `kivy`.

Leaving `Cython` in install-requirements, would result in install
problems as `kivy` would be buildt first.