Disable setting of LC_* env vars with SSH.
SSH normally allows setting LC_LANG vars when sshing into a host. On
the raspberry pi this breaks the virtual keyboard due to a complicate
cascade of settings and shortcomings in Python, kivy and default
Debian settings.

We therefore forbid setting LC_* vars and can trust that the default
en_GB.utf-8 setting is active when running `waeup_identifier`.