Remove test locations from pytest `addopts`.
The `addopts` option always adds the given paths when running
tests. We therefore can not run certain single test files even if
specified on the commandline.

Using the `testpaths` option in `tox.ini` we get this possibility
back.

The `testpaths` option was introduced in pytest 2.8. We will have to
update our dependencies.