Extend `fake_kofa_server` to populate its db.
The fake_kofa_server provides a kofa-like webservice XMLRPC
interface. The API is a bit extended to allow debugging and the like,
but mainly it serves as a testing replacement where no real Kofa
server should or can be used.

With the new `-p` option we allow to create student entries
automatically in the (normally empty) fake database.