Catch all exceptions in webservice functions.
Yet we catched xmlrpc-related exceptions only. Socket errors and
the like were passed on unhandled. Since it probably makes no
sense to catch some types of exceptions but others not, we now
catch everything and turn it in an error string.

This might be justified by the fact that the functions are normally
run in hard-to-control concurrent threads.