Hardcode version number in sphinx-setup.
To retrieve the version number from setuptools/pkg_resources, we have to
install the complete package before. While this would be okay for real
installs, it is too much for docs-only builds as run on readthedocs.org.

We now have to take care, that the version number set in `setup.py` is
also set in docs `conf.py` when changed.