Set raspi setup vars as facts.
This way the vars are also visible in included playbooks, for instance
the ssh setup playbook. This is important in order not to confuse users
that want to set, for instance, `reverse_ssh_port`. A variable of this
name is set in `setup_raspi_playbook.yml` and also in
`setup_ssh_playbook.yml`. But setting the var in the former will not
change the setting in the latter, although it is included by the former.
With using `set_fact` we make these vars 'global' and set it in all
included playbooks simulatnously. Just the behaviour one would expect
when running playbooks including other playbooks, I think.