Add ansible task for removal of [input] setting.
Should fix double output when using an external keyboard. When using the
old config, key presses of an external keyboard are output two times.
As the config installed before was tested (and worked correctly),
chances are, that this problem was introduced with some system changes
in Raspbian. Frankly, I do not know. But removing the one line in the
[input] section of kivys config.ini seems to fix the problem.