Wait for keypress at end.
When running on raspberrypi, we start this script in a terminal
window but the terminal will normally close immediately, while
the script runs on.

Using the keypress read, the terminal window remains visible.

The keypress is pureley cosmetical. The reverse tunnel built
will not be torn down when the terminal vanishes.