Finetune maintenance box setup.
We will now

- create backups of the sshd config when modifying it
- forbid password auth (keys only) for reverse user
- forbid TCP- and X11 forwarding
- use our own mark (enable other ansible playbooks to work additionally)
- reenable global scope of rules at end ("Match all").

Without the last step, any settings done in sshd config following our
stanzas would be  valid for the reverse user only. This is because a rule
scope set with "Match" is valid for all following settings or until the next
"Match" appears. With "Match all" we reset the rule scope back to the
default for following rules.