Add task to enable single click mode.
As double-clicking app launchers on the desktop became somewhat
difficult with PIXEL, we try single-click launching of apps. This is
driven by the filemanager settings as the desktop (as far as app
launching is concerned) seems in fact to be a filemanager application.

This might be the better solution compared to putting the app launchers
into the upper desktop bar.

Give it at least a try.